feat(Separator): Separator 객체 생성에 따른 StringParser 책임 분리
-  아래 6개 메소드 Separator객체로 이동
  splitByAllSeparators
  extractCustomSeparator
  removeCustomSeparatorDefinition
  findCustomSeparatorEndIndex
  addCustomSeparator
  createSeparatorRegex
